feat: Stock and Transaction View

+++ api/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render
 from rest_framework import status
-# Create your views here.
+from rest_framework.authentication import BasicAuthentication
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
@@ -8,34 +9,58 @@ from api.serializers import *
 
 
 class StockView(APIView):
-    def get_object(self, pk):
-        try:
-            return Stock.objects.get(pk=pk)
+    def get(self, request, stock_id=None):
+        if stock_id is not None:
+            stock = self.get_object(stock_id)
+            if stock:
+                serializer = StockSerializer(stock)
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response(data={"msg": "Stock not found"}, status=status.HTTP_404_NOT_FOUND)
 
-        except:
-            raise status.HTTP_404_NOT_FOUND
-
-    def get(self, request):
         stocks = Stock.objects.all()
         serializer = StockSerializer(stocks, many=True)
-        return Response(serializer.data)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
         serializer = StockSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.erros, status=status.HTTP_400_BAD_REQUEST)
-
-    def put(self, request, pk):
-        stock = self.get_object(pk)
-        serializer = StockSerializer(stock, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, pk):
-        stock = self.get_object(pk)
-        stock.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+    def get_object(self, stock_id):
+        try:
+            return Stock.objects.get(id=stock_id)
+        except Stock.DoesNotExist:
+            return None
+
+    def put(self, request, stock_id):
+        stock = self.get_object(stock_id)
+        if stock:
+            serializer = StockSerializer(stock, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response("Stock not found", status=status.HTTP_404_NOT_FOUND)
+
+    def delete(self, request, stock_id):
+        stock = self.get_object(stock_id)
+        if stock:
+            stock.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        return Response("Stock not found", status=status.HTTP_404_NOT_FOUND)
+
+
+class TransactionView(APIView):
+    def get(self, request):
+        transactions = Transaction.objects.all()
+        serializer = TransactionSerializer(transactions, many=True)
+        return Response(serializer.data)
+
+    def post(self, request):
+        serializer =  TransactionSerializer(data=request.data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
